Usernames => Emails
Switch to using emails as usernames. Because anonymity isn't a concern
for this app, it makes no sense to hide behind an obscure username. Not
to mention, if they really want to they can simply change their display
name. Not to mention, this builds in a required contact point for us if
we ever need to reach out to users.